KAFKA-12932: Interfaces for SnapshotReader and SnapshotWriter (#11529)
Change the snapshot API so that SnapshotWriter and SnapshotReader are interfaces. Change the existing types SnapshotWriter and SnapshotReader to use a different name and to implement the interfaces introduced by this commit.

Reviews: José Armando García Sancio <jsancio@users.noreply.github.com>
