+++ metadata/src/test/java/org/apache/kafka/controller/QuorumControllerTest.java
@@ -77,8 +77,9 @@ import org.apache.kafka.metadata.RecordTestUtils;
 import org.apache.kafka.metalog.LocalLogManagerTestEnv;
 import org.apache.kafka.raft.Batch;
 import org.apache.kafka.server.common.ApiMessageAndVersion;
-import org.apache.kafka.snapshot.RawSnapshotReader;
 import org.apache.kafka.snapshot.SnapshotReader;
+import org.apache.kafka.snapshot.RawSnapshotReader;
+import org.apache.kafka.snapshot.RecordsSnapshotReader;
 import org.apache.kafka.test.TestUtils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
@@ -493,7 +494,7 @@ public class QuorumControllerTest {
     }
 
     private SnapshotReader<ApiMessageAndVersion> createSnapshotReader(RawSnapshotReader reader) {
-        return SnapshotReader.of(
+        return RecordsSnapshotReader.of(
             reader,
             new MetadataRecordSerde(),
             BufferSupplier.create(),
